feat: added the bayes function
Wrote a function called: bayes that accepts the pior probabilities
that were provided by the user via the command line. The joint,
normalizer and then the posterior probability are calculated within
the bayes function and the posterior is returned.

The returned value is output in the main function, which is also
the calling function.